give wind stuff its own file
